refactor
Headers, required by implementation of a particular module, are included in its source instead of header.
(dst,src)-styled parameter list of encoder/decoder.
bitio.h - 64bit version.

+++ bitio.h
@@ -5,21 +5,21 @@
 #include <stdlib.h>
 #include <stdint.h>
 
-#define UINTWID (sizeof(uint32_t) * 8)
+#define UINTWID (sizeof(uint64_t) * 8)
 
 typedef struct {
     FILE *file;
-    uint32_t ibuf, ipos,
+    uint64_t ibuf, ipos,
              obuf, opos;
 } bitio_t;
 
 #define BITIO_INIT {NULL, 0, UINTWID, 0, 0}
 
-inline int bget(bitio_t *this, uint32_t *pbits, uint32_t nbits)
+inline int bget(bitio_t *this, uint64_t *pbits, int nbits)
 {
     nbits = nbits <= UINTWID ? nbits : UINTWID;
     *pbits = (this->ipos < UINTWID)
-           ? (this->ibuf >> this->ipos) & ((uint32_t )-1 >> (UINTWID - nbits))
+           ? (this->ibuf >> this->ipos) & ((uint64_t )-1 >> (UINTWID - nbits))
            : 0;
     this->ipos += nbits;
     if (this->ipos <= UINTWID)
@@ -27,18 +27,18 @@ inline int bget(bitio_t *this, uint32_t *pbits, uint32_t nbits)
 
     this->ibuf = 0;
     this->ipos -= UINTWID;
-    uint32_t rcnt = 8 * fread(&this->ibuf, 1, sizeof(uint32_t), this->file);
+    int rcnt = 8 * fread(&this->ibuf, 1, sizeof(uint64_t), this->file);
     if (rcnt < this->ipos)
         return -1;
 
-    *pbits |= (this->ibuf & ((uint32_t )-1 >> (UINTWID - this->ipos)))
+    *pbits |= (this->ibuf & ((uint64_t )-1 >> (UINTWID - this->ipos)))
               << (nbits - this->ipos);
     this->ibuf <<= UINTWID - rcnt;
     this->ipos += UINTWID - rcnt;
         return 0;
 }
 
-inline void bput(bitio_t *this, uint32_t bits, uint32_t nbits)
+inline void bput(bitio_t *this, uint64_t bits, int nbits)
 {
     nbits = nbits <= UINTWID ? nbits : UINTWID;
     if (nbits != UINTWID) bits &= (1 << nbits) - 1;
@@ -47,7 +47,7 @@ inline void bput(bitio_t *this, uint32_t bits, uint32_t nbits)
     if (this->opos < UINTWID)
         return;
 
-    fwrite(&this->obuf, sizeof(uint32_t), 1, this->file);
+    fwrite(&this->obuf, sizeof(uint64_t), 1, this->file);
     this->opos -= UINTWID;
     this->obuf = this->opos ? bits >> (nbits - this->opos) : 0;
 }
